third_party/WebKit/LayoutTests/fragmentation/margin-bottom-at-top-of-fragmentainer.html

<!DOCTYPE html>
<p>There should be a blue square below.</p>
<div style="position:relative; columns:2; column-gap:0; column-fill:auto; width:200px; height:100px;">
    <div style="margin-bottom:50px; height:100px;"></div>
    <div id="elm" style="height:100px; background:blue;"></div>
</div>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(() => {
    var elm = document.getElementById("elm");
    assert_equals(elm.offsetLeft, 100);
    assert_equals(elm.offsetTop, 0);
}, "Margins should collapse with fragmentainer boundaries");
</script>
